fix: cap UTF8Encoding internal array size to 1MiB
When decoding large strings with the UTF8Encoding, it will expand the
backing array for future use. Unfortunately, when combined with a
connection pool, the retained buffer can cause high memory usage and a
potential OutOfMemoryError. Cap the the internal buffer to 1MiB to avoid
this problem.

Larger strings will use a new array allocation.